+++ R/dists.R
@@ -1,4 +1,5 @@
 
 
 survstan_distributions <- c("exponential", "weibull", "lognormal", "loglogistic",
-                            "fatigue", "gamma", "rayleigh")
+                            "fatigue", "gamma", "rayleigh", "gompertz",
+                            "ggstacy", "ggprentice")