+++ NEWS.md
@@ -48,3 +48,9 @@
 # survstan 0.0.6.1
 
 - Correcting parametrization of survival functions for AH and EH models.
+
+
+# survstan 0.0.7
+
+- All implemented models now can accommodate a offset variable in the linear predictors.
+- Implementation of Gompertz, generalized gamma (original Stacy's parametrization and alternative Prentice's parametrization) distributions.